Adding Scraping Controller and Route
