+++ conf/routes
@@ -6,5 +6,8 @@
 # An example controller showing a sample home page
 GET     /                           controllers.HomeController.index()
 
+# Route for turning on scraping job
+GET /scrape                         controllers.ScraperController.index()
+
 # Map static resources from the /public folder to the /assets URL path
 GET     /assets/*file               controllers.Assets.versioned(path="/public", file: Asset)
